HW-9 tabsFragment 코드 정리 및 databinding 오류나는 문제 수정
