+++ 2giwon/ArchitectureStudy/app/src/main/res/layout/fg_tab.xml
